Move smooth code to hook
